refactor(stella-context): split domain scoping into retrieval::scope
`RecallScope` pushed `retrieval.rs` to 1519 lines, past the 1500-line ratchet,
and the file is not grandfathered — so it gets split, not baselined
(AGENTS.md § "God files"). 1444 lines now, with room.

The split is along a real seam rather than wherever the line count fell.
`retrieval::scope` owns everything about domain scoping: the two-scope type
and the two projections the corpus tag map feeds — `overlap_ranking` (which
ranks, against the session scope) and `evidence_ids` (which admits, against
the query scope). Putting them side by side is the point: they read the same
map through different scopes, the asymmetry is the subtlest thing in
retrieval, and it has already been got wrong once (#2289).

Both are pure functions over already-loaded rows, matching the idiom of the
sibling modules `evidence` and `ranking`; the parent keeps the one query that
loads the map.
